# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Пользователь успешно создан",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserWithoutRoles"
                }
              }
            }
          },
          "400": {
            "description": "Некорректные данные"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Создание нового пользователя",
        "tags": [
          "user"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "страница",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "отображаемых записей на странице",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "поле для сортировки",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": false,
            "in": "query",
            "description": "порядок сортировки ASC/DESC",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "поиск",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filters",
            "required": false,
            "in": "query",
            "description": "фильтры (объект в виде JSON-строки)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список пользователей успешно получен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "403": {
            "description": "Нет доступа"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение списка всех пользователей",
        "tags": [
          "user"
        ]
      }
    },
    "/user/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Информация о себе успешно получена",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение информации о себе",
        "tags": [
          "user"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID пользователя",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно найден",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Получение пользователя по ID",
        "tags": [
          "user"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID пользователя",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Данные пользователя успешно обновлены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Обновление данных пользователя",
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID пользователя",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Пользователь успешно удален",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/"
                }
              }
            }
          },
          "404": {
            "description": "Пользователь не найден"
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Удаление пользователя",
        "tags": [
          "user"
        ]
      }
    },
    "/user/set-roles/{id}": {
      "post": {
        "operationId": "UserController_setRoleToUser",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID пользователя",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "roles": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "description": "Массив названий ролей",
                    "example": [
                      "admin",
                      "user"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Роли успешно назначены пользователю"
          },
          "404": {
            "description": "Пользователь или роли не найдены",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Назначение ролей пользователю",
        "tags": [
          "user"
        ]
      }
    },
    "/user/change-password": {
      "post": {
        "operationId": "UserController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пароль успешно изменен",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserChangePasswordResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Смена пароля",
        "tags": [
          "user"
        ]
      }
    },
...
  },
  "info": {
    "title": "Passport API",
    "description": "API документация для системы управления пользователями, ролями и разрешениями\n\nСпецификация OpenAPI доступна в формате JSON: [openapi.json](/api/openapi.json)",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "john_doe",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Уникальный email пользователя",
            "example": "john_doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "password123",
            "minLength": 6
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя пользователя",
            "example": "John Doe"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя",
            "example": "https://example.com/avatar.jpg"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона пользователя",
            "example": "+1234567890"
          },
          "authProvider": {
            "type": "string",
            "description": "Провайдер аутентификации (например, local, google, github)",
            "example": "local"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя (например, en, ru)",
            "example": "en"
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс пользователя",
            "example": "UTC"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "UserWithoutRoles": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя (уникальное)"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя (уникальный)"
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "writeOnly": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Статус активности пользователя",
            "default": true
          },
          "isVerified": {
            "type": "boolean",
            "description": "Статус верификации пользователя",
            "default": false
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона пользователя"
          },
          "authProvider": {
            "type": "string",
            "description": "Провайдер аутентификации (например, local, google, github)"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя (например, en, ru)"
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс пользователя"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "password",
          "isActive",
          "isVerified"
        ]
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Список пользователей",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserWithoutRoles"
            }
          },
          "total": {
            "type": "number",
            "example": 100,
            "description": "Общее количество записей"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
      "UserMeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID пользователя"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя"
          },
          "roles": {
            "description": "Роли пользователя",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "username",
          "roles"
        ]
      },
      "RoleResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Название роли (уникальное)"
          }
        },
        "required": [
          "name"
        ]
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Уникальный идентификатор пользователя"
          },
          "username": {
            "type": "string",
            "description": "Имя пользователя (уникальное)"
          },
          "email": {
            "type": "string",
            "description": "Email пользователя (уникальный)"
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя пользователя"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "writeOnly": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Статус активности пользователя",
            "default": true
          },
          "isVerified": {
            "type": "boolean",
            "description": "Статус верификации пользователя",
            "default": false
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона пользователя"
          },
          "authProvider": {
            "type": "string",
            "description": "Провайдер аутентификации (например, local, google, github)"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя (например, en, ru)"
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс пользователя"
          },
          "roles": {
            "description": "Роли пользователя",
            "example": [
              {
                "name": "admin"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponse"
            }
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "password",
          "isActive",
          "isVerified",
          "roles"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Имя пользователя",
            "example": "john_doe",
            "minLength": 3
          },
          "email": {
            "type": "string",
            "description": "Уникальный email пользователя",
            "example": "john_doe@example.com"
          },
          "password": {
            "type": "string",
            "description": "Пароль пользователя",
            "example": "password123",
            "minLength": 6
          },
          "fullName": {
            "type": "string",
            "description": "Полное имя пользователя",
            "example": "John Doe"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL аватара пользователя",
            "example": "https://example.com/avatar.jpg"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Номер телефона пользователя",
            "example": "+1234567890"
          },
          "authProvider": {
            "type": "string",
            "description": "Провайдер аутентификации (например, local, google, github)",
            "example": "local"
          },
          "locale": {
            "type": "string",
            "description": "Локаль пользователя (например, en, ru)",
            "example": "en"
          },
          "timezone": {
            "type": "string",
            "description": "Часовой пояс пользователя",
            "example": "UTC"
          },
          "isActive": {
            "type": "boolean",
            "description": "Статус активности пользователя",
            "example": true
          },
          "isVerified": {
            "type": "boolean",
            "description": "Статус верификации пользователя",
            "example": true
          }
        }
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "Новый пароль",
            "example": "newPassword123",
            "minLength": 6
          },
          "previousPassword": {
            "type": "string",
            "description": "Старый пароль",
            "example": "newPassword123",
            "minLength": 6
          }
        },
        "required": [
          "newPassword",
          "previousPassword"
        ]
      },
      "UserChangePasswordResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение о результате операции",
            "example": "Пароль успешно изменен"
          }
        },
        "required": [
          "message"
        ]
      },
...
    }
  }
}